# optimization: precalculate SipHash constant XOR with k0 and k1 in SaltedOutpointHasher