--- a/src/crypto/siphash.h
+++ b/src/crypto/siphash.h
@@ -37,18 +37,15 @@ public:
     uint64_t Finalize() const;
 };
 
-/** Optimized SipHash-2-4 implementation for uint256.
+/**
+ * Optimized SipHash-2-4 implementation for uint256.
  *
- *  It is identical to:
- *    SipHasher(k0, k1)
- *      .Write(val.GetUint64(0))
- *      .Write(val.GetUint64(1))
- *      .Write(val.GetUint64(2))
- *      .Write(val.GetUint64(3))
- *      .Finalize()
+ * This class caches the initial SipHash v[0..3] state derived from (k0, k1)
+ * and implements a specialized hashing path for uint256 values, with or
+ * without an extra 32-bit word. The internal state is immutable, so
+ * PresaltedSipHasher instances can be reused for multiple hashes with the
+ * same key.
  */
-uint64_t SipHashUint256(uint64_t k0, uint64_t k1, const uint256& val);
-
 class PresaltedSipHasher
 {
     uint64_t v[4];
@@ -61,6 +58,7 @@ public:
         v[3] = CSipHasher::C3 ^ k1;
     }
 
+    uint64_t operator()(const uint256& val) const noexcept;
     uint64_t operator()(const uint256& val, uint32_t extra) const noexcept;
 };