@paplorinc 2024-07-14
That's disappointing, thanks for checking, something's still off with my compiler, it seems.
After I finish benching https://github.com/bitcoin/bitcoin/pull/28280#issuecomment-2227314032, I'll try this again on the dedicates server instead.
I'm also a bit surprised that the ratios are completely different compared to my measurements, e.g. `SaltedOutpointHasherBenchmark_match` and `_mismatch` are 40% off, while in my case they're basically equal. Are compilers this different?
Could you please send me the exact command you've used and the compiler versions you've used?
Thanks!