@andrewtoth 2024-07-15
>  Could you please send me the exact command and compiler versions you've used?

`gcc version 11.4.0 (Ubuntu 11.4.0-1ubuntu1~22.04) `

I ra
`./configure --enable-bench`
`make`
`./src/bench/bench_bitcoin -filter=.*Out[pP]oint.*`